[FP-76] Refactor: types폴더 type파일 사용하는 폴더 및 파일로 이동 후 정리
