--- src/types/sideNavMenuTypes.ts
@@ -1,5 +1,0 @@
-export default interface SideNavMenuOptionProps {
-  imgSrc: JSX.Element;
-  text: string;
-  isActive: boolean;
-}
